+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Versioning: [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
 
-## [Unreleased]
+## [0.15.0] - 2026-05-23
 
 ### Changed